Move templates to central location, fix packr build
This change does a couple of things:

1. Adds windows and linux build artifacts
2. Fixes the Packr build so static files are actually in
   the binary
3. Moves all of our templates to a central templates package

My makefile-fu is weak, but I've tested the binaries on a couple
of different linux machines and they're good. Also got the unit
tests to fail when the binary isn't compiled with packr artifacts.

The way packr works is it generates go files with the static files
as bytes, compiles your code against those generated .go files,
and then deletes the generated files from your workspace.