Use different logic to determine GIT_DIRTY
Previously, we were checking if a branch was dirty based on whether or
not there was output when `git diff` was called. If there was no diff,
$? would equal 0 and we based prefixing the git short hash with '*'
accordingly. However, it would be more accurate if we based the logic on
whether git HEAD has been tagged (presumably, with a version, though
admittedly this would break if we happened to tag where we were locally
and ran `make`.) This is probably unlikely however, assuming that our
release process involves tagging commits with the version only.