Attach environment principal to DNS Delegation Policy
This change makes it so we can do cross-account DNS delegation. It
grants the environment permission to update the project's hosted zone.

TODO: the project.yml template is not complete, it just adds a role.

But this change adds the infrastructure to add an environment to a
project's DNSDelegation role.