Fix Multiple Listner Rules
Right now, we pass in the listner rule value of 1.
This works for 1 app, but not if we want to add another.
To solve this, we have to use a custom resource to generate
the next available rule priority by looking at the existing LB
listner rules, and returning the max + 1.

This change also fixes a race condition with spinning up the
listner and the ECS service (the sevice needs the listner
to be created first). I had to do a little trick do to a conditional
depends on using a waiter.

This change, completly unrelated, also exposes the LB DNS path
as an env variable to the task definition we create.