Replace ChangeSet suffix with UUID
This change lets us recover from empty stack updates.
Right now, if we generate a changeset, we suffix it
with the image tag.

If this is empty, the changeset is kept in CF.
If we make a change that doesn't affect the tag,
we'll fail to deploy.

We should clean up those failed changesets, but for now,
to get this working, changing the ID to be a UUID.

addresses #460