Fix unit test failures when ecs-project folder is in context
Addresses #332 where unit tests will fail if there is an ecs-project folder with a .ecs-workspace file which contains a project name.
The unit tests were picking that value up via the cli package init func which in turn causes unit tests that rely on an empty project name to fail.