Remove unused method
This was leftover from a refactor in which we moved much of the PreRunE
logic into the constructor for the opts. This approach allows for ease
and consistency of testing.

NOTE: See 16d3bb1 -- previously, calling the constructors outside of
RunE made it harder to catch errors returned by the constructor itself.
This allows the errors to be caught will still getting values from the
command's flags..