test: First Pass at E2E tests
This is my first pass at real end to end tests. There are currently
3 suites:

1. Init Flow
  This is our simple, one command and have a service up command test,
  and ensures that an app can start up, logs, ls, delete work and
  the public endpoint resolves.
2. Multi Env Project
   This sets up a project with two environments, in different accounts
   and regions. It deploys an app to both environments, ensures that
   the app is reachable, and the following commands work:
      app [show, logs, ls, deploy, init],
      project [show, ls, init, delete],
      env [init, ls]
3. Multi App Project (single ws)
   This tests a similar set of commands, but ensures that we can deploy
   two apps to one environment. The app reachability test (calling the http
   endpoint) currently fails because that feature is broken without updating
   the manifest.

There are lots of todos, but wanted to get this out as it's a good start:
* App Package
* Pipeline init, update, delete
* Pipeline release (git push and wait for something to go out)
* Multi workspace, multi app project

Another important note - this change doesn't currently wire
the e2e tests into our release process. That'll be a different
PR. These tests can run in a Docker image so that we don't have
to worry about dependencies, filesystem cleanup, etc.