feat: add profile flag to root init command (#642)
* feat: Add profile flag to init command

This commit adds a profile flag to `ecs-preview init` which is passed
into the implicit call to `ecs-preview env init`. If the user does not
specify a profile, ask them to select a profile.

Furthermore, `env init` is refactored to only initialize session clients after the user is prompted for a profile, eliminating a bug where we only deployed new environments to the user's default profile.