test(e2e): add some buffer for validating endpoints
There seems to be times when we call the endpoint, but
it's not quite ready yet, after we do a deployment. I can't
reproduce this localy, which leads me to believe it's because
CodeBuild instance is just faster.

This change lets the tests retry a few times (up to 10s) before
failing.