feat!: Use / as the default path for new apps
Currently, each Load Balanced Web App is created and is asigned
a path of "/{app name}" on the ALB. This makes sense, but there
are some gotchas.

First, your app has to be prepared to handle paths of "/{app name}".
This is probably a little unlikely unless your lucky. Most apps
expect to be run as if they are receiving requests on the root
path "/" (see every tutorial ever) 😂

What this change does is it defaults your first app to being hosted
on "/" but at the last evaluated rule. This is probably a more
intuitive default (and we've gotten some customer feedback on this).

Subsequent apps still have to use the path based route.

So:

```sh
ecs init --app front-end # produces a URL of http://my-lb.us-west-2.amazon.com/
ecs init --app api       # produces a URL of http://my-lb.us-west-2.amazon.com/api/
```

This change adjusts the priority generator so that "/" is evaluated last.

Customers are still free to change the path in their manifest. They can use
the special "/" path to signify that they want this to be their default route.

__ Why didn't you just use the default action ? __

One kind of odd thing we did here is give the "/" path a really high rule
priority, rather than just setting it as the default route. There are a few
reasons we did that.

1. It's a bit messy as the default route action is managed in the environment stack
   and there's really not a great way to update it.
2. If a customer changes their app to use a named path, rather than the default "/"
   path - it will only require a dpeloyment of the app, rather than then the app
   AND the env (which is a different stack).

This change also has some other E2E bug fixes in it (like setting the healthcheck
path by default).