refactor: delete project opts
This change adds a `sessionProvider` field and uses `getBucketEmptier` instead
of having `initBucketEmptySvc` return a mutable `s3Client` field on the
DeleteProjectOpts struct, which makes the s3Client easier to mock during unit
testing.

Also got rid of the `projectResourcesGetter` field since this interface is
embedded in the deployer interface.