test: happy path for DeleteProjectOpts#Execute
This involved major refactoring in the DeleteProjectOpts to allow
mocking of deleting other resources such as Apps, Envs, and Pipelines,
as these involve instantiating the respective Delete*Opts and calling
the main Execution path, which also involve a lot of calls that would
need to be mocked. This abstracts the entire Exeuction call into
effectively an empty method in an interface that we can simply succeed
or return an error.