feat: pipeline show validate method (#821)
* add `store` field to ShowPipelineOpts

This interface provides a way to make calls to SSM, which is where the
names of the project (as well as envs and apps) are stored.

* validate that project name exists

* validate that pipeline exists

* implement pipeline_show#Ask, part the first

Move the logic of reading the pipeline manifest into Ask instead of
Validate. Validating the pipeline name will now involve making a call to
codePipeline (TODO)

* add pipelineGetter interface and codepipeline client

* validate pipeline name for real this time

* chore: add tests and cleanup