fix: do not fail project delete if pipeline manifest does not exist (#836)
* chore: move workspaceDeleter into interfaces.go

This allows for easier mock generation and reusability

* refactor: delete project opts

This change adds a `sessionProvider` field and uses `getBucketEmptier` instead
of having `initBucketEmptySvc` return a mutable `s3Client` field on the
DeleteProjectOpts struct, which makes the s3Client easier to mock during unit
testing.

Also got rid of the `projectResourcesGetter` field since this interface is
embedded in the deployer interface.

* test: happy path for DeleteProjectOpts#Execute

This involved major refactoring in the DeleteProjectOpts to allow
mocking of deleting other resources such as Apps, Envs, and Pipelines,
as these involve instantiating the respective Delete*Opts and calling
the main Execution path, which also involve a lot of calls that would
need to be mocked. This abstracts the entire Exeuction call into
effectively an empty method in an interface that we can simply succeed
or return an error.

* fix: do not fail project delete if no pipeline manifest

* fix: validate pipeline manifest existence in pipeline delete

* chore: rename interfaces

to appease the style gods: https://golang.org/doc/effective_go.html#interface-names

* refactor: change ErrNoPipelineInWorkspace to error

* chore: remove 'fmt' prefix for progress message constants

* refactor: set envProfile within DeleteEnvVars