chore: create config pkg to hold app config
This change copies the store functionality into a new package
called config. It also copies the archer structs for App, Env, Service
to this config package.

We can slowly start removing the archer package and now use
config.Application, config.Environment, config.Service instead.

Leaving the old package so we don't break everything that depends
on it, but we can consider it deprecated and start replacing it
with this package.