fix: cicd buildspec works with copilot binary (#957)
Since the name of the copied binary is "copilot" and
the name of the directory where we put our manifest files
is also "copilot", running "./copilot" would result in trying
to execute the directory and the build stage failed.

This change updates the copied binary name to "copilot-linux"
to avoid the confusion.