feat(prompt): Add multiselect prompter option  (#1013)
<!-- Provide summary of changes -->
Adds a multiselect based on survey.MultiSelect to the list of prompter methods. This will be used when selecting which attributes to use for extra LSI sort keys on a DDB table. 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.