fix(storage): Remove "attribute" concept from UX (#1037)
<!-- Provide summary of changes -->
This PR removes the notion of an Attribute from the UI of `copilot storage init`. 

Previously, customers could specify extra attributes via the `--att` flag which they could specify for use in Local Secondary Indices. This is problematic, however, because DynamoDB does not allow unused attributes to be written in the CF template and this would cause the addons stack deployment to fail. 

Removing "attribute" as a customer facing concept simplifies the user experience by only asking for "alternate sort keys/LSIs" and allows us to avoid complex validation issues which arise when extra attributes are specified. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.