fix(cli): Correctly set NoLSI field when askDynamoLSIConfig returns false (#1057)
<!-- Provide summary of changes -->
This change solves an issue where if a customer selected "No" when asked "Would you like to add any alternate sort keys to this DynamoDB table?", the `noLsi` option was not properly set and an erroneous line was injected into the CF template. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.