chore(pipeline): add actions to pipeline status (#1051)
This change enables the `pipeline status` subcommand to display status information about each action within each stage (in addition to the existing status information about each stage). 

(The first commit singles out TestCommands and does not show the other commands, if that is preferable. I believe info on all of the stages is useful for a full picture of a pipeline's status.)

Resolves https://github.com/aws/amazon-ecs-cli-v2/issues/1045

Tree-style output to be implemented in a separate PR. https://github.com/aws/amazon-ecs-cli-v2/issues/1052

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.