chore: add color to pipeline status (#1078)
These changes draw attention to some of the output of `pipeline status`:
"Failed" is red
"InProgress" is bold/emphasized
"DISABLED" is faint

This is a work-around for https://github.com/aws/copilot-cli/issues/981 and also the beginnings of a solution for https://github.com/aws/copilot-cli/issues/1071. 

The swapping of the Status and Transition columns may help suggest to users that "Transition" refers to whether the _incoming_ transition is enabled or disabled, and also eliminates the problematic tab after the color-added words. Tabwriter doesn't work well with ANSI formatting because of the control characters; apparently a fix is in progress: https://go-review.googlesource.com/c/go/+/202257/. Hopefully that problem is solved before we have to add another column to the right!


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.