fix: Display correct GitHub URL flag name for 'copilot pipeline init' command (#1114)
When calling `copilot pipeline init`, if Copilot cannot parse the GitHub repo from the provided URL an error message is thrown. There is a reference to a CLI flag, `--url`, which seems to be incorrect. The correct flag is `--github-url`. This PR changes it.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.