chore(cli): update image to task for `task run` (#1183)
This PR builds and pushes image and updates image to task definition for `task run`. This is a part of the series of PRs that implement `Execute` for `task run`. 

Related #702 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.