chore(task): modify return values for `Run` in task package (#1198)
This PR contains the following changes:
1. refactor ecs pkg: modify `RunTask` to return a list of `Task` struct instead of `taskARNs`
2. refactor task pkg: modify `Run` to return a list of `Task` struct wrapped by the `task` pkg instead of `taskARNs`

In some use cases, `taskARNs` are not the only information that the caller need. For example, a caller that run tasks and then display logs may also need the `StartedAt` information to filter the logs. 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.