fix: add UntagResource permission for KMS key (#1206)
There seems to be an issue with the current KMS CF resource
that is causing superfluous tag/untag calls. This new permission
fixes that issue.

A customer encountered this issue (#1199) when using resource tags,
so I've also updated the e2e tests to include setting resource tags.

In the future, this permission will be useful if we want to support
updating tags at the app level (removing a tag, for example) so it's
not entierly a hack :)

Fixes #1199 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.