chore(ecs): return container stopped reason from RunTask (#1237)
This PR refactors `RunTask` method from `ecs` package to return more informative error message. 

#### Why 
The error message returned from the waiter is too generic. For example, currently the error message eventually returned to the user may look like this:
```
✘ run task test: run task test: wait for tasks to be running: ResourceNotReady: failed waiting for successful resource state
```


#### How
When the waiter fails to wait for the tasks to start, if it's a `ResourceNotReady` error, we try to return the container-level failure reason, which is more informative. For example, the error message may look like this:
```
✘ run task test: run task test: task 590bc4fe: CannotStartContainerError: Error response from daemon: OCI runtime create failed: container_linux.go:348: starting container process caused "exec: \"very-random-command\": executable file not found in $PATH": unknown
```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.