chore: Lower Service Discovery DNS TTL
This lowers the service discovery down from 1min to 10s. This value is how long a IP address is cached. While we wait for https://github.com/aws/containers-roadmap/issues/343 to be resolved, we can lower the value to 10seconds. The longer this value is, the longer a service might fail to call another service due to a bad cache value.

fixes #1173