chore: lower Service Discovery DNS TTL (#1282)
This lowers the service discovery down from 1min to 10s. This value is how long a IP address is cached. While we wait for https://github.com/aws/containers-roadmap/issues/343 to be resolved, we can lower the value to 10seconds. The longer this value is, the longer a service might fail to call another service due to a bad cache value.

fixes #1173

<!-- Provide summary of changes -->

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.