docs(site): add document for "task run" command. (#1344)
Wiki page has the document for `copilot task run` command, but there is no file related to it in this repository. Adding markdown based on wiki's one.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.