refactor(cloudformation): rename ServiceName parameter to WorkloadName (#1340)
Renames `ServiceName` and associated constructs to `WorkloadName`. This is not a breaking change, as it does not change SSM parameters. 

It does, however, change the names of some internal structs for better compatibility with the upcoming scheduled task work. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.