feat(autoscaling): override desired count to autoscaling count during deployments (#1371)
Part of #1349. Add custom resource lambda function so that we can delegate desired count smartly. By default when a service is deployed for the first time, we take either the desired count specify by the user or the minimum value of their auto scaling range (if exists) as the ECS service desired count. When they update their service, we detect their current running copies and set the ECS service desired count to remain the same, so as to avoid users having this issue https://github.com/aws/containers-roadmap/issues/493.

<!-- Provide summary of changes -->

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.