refactor(workspace): enable distinguishing between manifests of service type and job type (#1386)
This adds JobNames() to ServiceNames() in the workspace package. These methods, with the aid of isJob() and isService(), which were previously housed in the manifest package, return a list of jobs/services selected by checking the `type:` field of the manifest yaml. 

Related #1131 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.