fix(cli): --app flag can now be overridden by users
Resolves #1329

Removes GlobalOpts struct and the viper lib dependency from the CLI.

Previously, if we ran "env init --app other" the overridden value
"other" was ignored and only the workspace's app was used.
This is an issue with viper as was found in #1329. Instead, with this
change we get rid of Global Flags and force commands themselves to write
whether they need an --app flag or not.