chore(cli): add Ask() method and unit tests to job_deploy subcommand (#1408)
This adds the Ask() method to the `job_deploy` subcommand. It is based on Ask() in `svc_deploy`, but includes a slight refactor in the selector package. 

Related #1131

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.