refactor(selector): move Dockerfile() and Schedule() into selector package (#1401)
<!-- Provide summary of changes -->

Moves the prompting logic for Dockerfiles and job schedules into the selector package so the CLI package doesn't have to handle it. 

related #1390 and #1131 : Prerequisite to job init being complete. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

Work in progress

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.