feat!(cli): delete profile flag from app and env delete cmds (#1510)
### Goal
Simplify the `env delete` and `app delete` commands so that a named profile is no longer required during deletion.  

Resolves #1420

### How
We now use the existing `CFNExecRole` to initiate a `DeleteStack` operation on the environment stack. However, since the role itself is defined in the environment stack that means that during deletion both the `CFNExecRole` and the `EnvManagerRole` should be retained. Therefore, the environment stack template is now updated to retain both roles and their deletions is handled separately within the CLI.


_By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice._