chore(cli): extract misleading language in job manifest for rates (#1554)
<!-- Provide summary of changes -->
At some point, we decided not to accept simple duration strings like "1h30m" in favor of the "@every 1h30m" syntax. 

This PR rewords the comment in the manifest to clarify the required syntax. 
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.