chore: public artifacts without version suffix in our releases
To link to latest releases in GitHub (see
https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/linking-to-releases)

We need to have a static name for our binaries that does not have the
version number in it.
With this change, we are now publishing the binaries as well as the asc
file without a version number.

The change is done within the signing stage because adding the new
binaries in the staging stage would possibly result in the Lambda to
timeout while signing.