chore: upgrade survey to v2.2.2 (#1651)
<!-- Provide summary of changes -->
We'll keep the current UI and update `survey` pkg to v2.2.2. The only UI change is when using `prompt.Get`, the cursor is hidden when entering the answer.
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.