chore(cli): override message when service name specified by flag (#1661)
This change recognizes if a single service in a workspace was specified by the `--name` flag, for a more sensible response. If it was not, the message `Only found one deployed service [svcName] in environment [envName]` is output; if the service was specified, it is not.

Affects `svc status` and `svc logs`. This behavior is now in line with that of `svc show`, `svc package`, `svc delete`.

Fixes https://github.com/aws/copilot-cli/issues/1416.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.