test(job): add jobs to e2e suites (#1652)
<!-- Provide summary of changes -->
This PR adds commands exercising scheduled jobs to the `init` and `multi-svc-app` suites. 
The "Init" test simply checks that a scheduled job is deployable through `copilot init`. 

The `multi-svc-app` suite now deploys a job, which uses service discovery to GET a route on the frontend service, allowing the test suite to verify that the job has executed. 

That is, the front-end service now has two more routes added to it:
* `/job-checker/`
* `/job-setter/`

A GET to `/job-setter/` changes the value of an environment variable in the frontend task from "no" to "yes". 
A GET to `/job-checker/` returns the value of the environment variable in the response body. 

The flow:
1. Frontend is deployed. 
2. Job is deployed. 
3. Other services are deployed. 
4. Within 4 minutes, the job executes, sending a GET to `http://front-end.$COPILOT_SERVICE_DISCOVERY_ENDPOINT/job-setter/`. This changes the value of the environment variable from "no" to "yes". 
5. After all services are deployed, the test suite sends a GET to `copilot-lb-endpoint.amazonaws.com/job-checker/`, verifying that the body of this response is "yes" and that the status code is `200`.
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.