chore(e2e): add healthcheck to backend e2e test (#1688)
This simply adds a health check to the backend Dockerfile. That the Dockerfile is correctly translated to the manifest is checked in manifest/svc_test.go. 

I ran this suite of tests locally and they passed. 

Addresses https://github.com/aws/copilot-cli/issues/1123.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.