fix(storage)!: allow attaching storage resources to jobs via `storage init` (#1701)
<!-- Provide summary of changes -->
This PR enables customers to select from the list of services AND jobs when creating a new addons template through Storage Init. 

It also replaces the `--svc/-s` flag with a generic `--for` flag, which may be a breaking change for some customers. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->
Addresses #1595 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.