feat(cli): display secrets as subcategory in `svc_show` (#1704)
This change displays secrets as a separate category (not within "Variables") in `svc show`. If the secret value comes from Parameter Store, it is preceded by `parameter/`. Otherwise, if it comes from Secrets Manager, it is displayed as an ARN.

Addresses https://github.com/aws/copilot-cli/issues/720

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.