chore: add validations to pipeline_init (#1759)
The one validation that was previously in Validate() was never reached-- users passing invalid (empty or nonexistent) appName via the `-a` flag got an `errNoEnvsInApp` because the constructor called getEnvs(), which threw that error. 

This adds appName validation within getEnvs() and also validates input for `-u` and `-e`. (Flag renaming-- `--github-url` -> `--url`-- will be in the next PR.)

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.