chore(command): add interactive run (#1761)
<!-- Provide summary of changes -->
Our `command.Run` doesn't work well if the command starts a child process, which allow users to run commands interactively: when the users press `ctrl+c`, the SIGINT will be passed to the parent process instead of the child process that users are interacting with. This PR adds a new method that allows interactive run: SIGINT signals can be received by the child process and ignored by the parent process, so that the parent process won't quit.
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.