fix(workspace): ignore unreadable directories when listing Dockerfiles (#1764)
<!-- Provide summary of changes -->
This is a quick change to skip over directories with invalid permissions when reading dockerfiles. If a directory doesn't have appropriate permissions, it's not essential that we read it--customers can always specify paths directly. 

This also fixes an issue introduced by the newest version of Survey (#1754) where the new OnInterrupt field was not populated, causing segfaults whenever ctrl+c was pressed at a prompt. 


<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.