chore: bump survey version and revert `OnInterrupt` shims (#1773)
<!-- Provide summary of changes -->

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->
The newest survey release reverts a commit which was introduced in the previous release, which added unstable behavior for all users of survey. 

This commit bumps to the latest version and reverts changes introduced in #1764 to handle the instability.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.