chore(cli): Unmarshal template in "env upgrade" to check if roles are retained
Previously, we would check if the following string was included in the
template to see if the EnvManagerRole or ExecRole were retained:
```yaml
EnvManagerRole:
    DeletionPolicy: Retain
```
Once, we added a comment under the EnvManagerRole:
```yaml
EnvManagerRole:
    # An IAM role to manage your environments
    DeletionPolicy: Retain`
```
The parsing broke and the CLI thought that the resources were not
retained.

Now, we're using the yaml.Unmarshal library to see if a "DeletionPolicy"
node is present under the roles.