chore(selector): add CFTask selector to list and select one-off task stacks (#1797)
<!-- Provide summary of changes -->
Adds a selector to get deployed one-off task stacks. 
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->
Related #1604. I'll close that PR in favor of this and my next one, which will contain the rest of the task delete logic. 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.