chore: add Done method to streamers and renderer to exit successfully
Previously, streamers and components from the `stream` and `progress`
packages respectively only exited if a `ctx` was canceled.

This PR enables `StackStreamer` to exit successfully on its own without
waiting for a context to cancel it when the stack is no longer in
progress.

Similarly, components that can both `Render` and `Listen` on events now
have a `Done` method to indicate that there are no more events to
`Listen` and we can exit successfully.