chore(cli): add task delete validate/ask (#1831)
<!-- Provide summary of changes -->
Adds the skeleton of a new `task delete` command with Ask() and Validate() filled in. Tests for Ask() forthcoming, then I'll work on Execute in a separate PR. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.