chore(ecs): refactor API for higher abstraction when stopping tasks
This commit refactors the API of the ecs package to allow the following
top level functions:
* StopDefaultClusterTasks
* StopAppEnvOneOffTasks
* StopActiveWorkloadTasks

All of these operate only on copilot tasks, whose families are of the form
of either `copilot-$task` or `$app-$env-$wkld`. These functions abstract the
logic of listing tasks in either the default or environment cluster and
stopping all instances.