chore(progress): fix showing cursor after rendering (#1901)
Although we have a `defer cursor.Show()` statement, it wouldn't actually get written to `out` because we never called `Flush()`. 

This change ensures that all buffered writes get written to `out` at the end of the invocation.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.