chore(template): add container paths as environment variables  (#1926)
Adds conditional environment variables based on volume definitions for a service or job. If there is a mounted volume `myEFSVolume` at path `/var/www`, it will be injected into the template as an environment variable. Additional volumes will be added to the JSON map. 
```yaml
- Name: COPILOT_MOUNT_POINTS
  Value: '{"myEFSVolume":"/var/www"}'
```


<!-- Provide summary of changes -->

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.