chore(cli): adapt tree to new treeprint version (#1963)
With the new version of treeprint, we are now able to specify the root of the tree. Previously, we had each stage as a branch as a workaround to avoid the extra dot and lines. This change restores the output of `pipeline status` to appear as it had, leveraging the new `NewWithRoot()`. Without this change, the stages appear as branches with extra lines.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.