chore(cli): prompt to select app if not under a workspace (#1958)
This pull requests changes current cli behavior for `env init`. If the user is not currently in any workspace and no application is specified using `--app` flag, it prompts the user to select from existing applications instead of returning an error. 

Resolves #1238  

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.