chore(stream): add exponential backoff and jitter to Fetch (#1935)
<!-- Provide summary of changes -->
this jithub pull request adds jitter and exponential backoff to the calls to Fetch in both Cloudformation and ECS streamers when they encounter a throttling exception. 

The new interval with jitter has an average duration of 2s, with exponential backoff for throttling exceptions up to 32s. It uses uniformly distributed random numbers between 0 and 4000ms for the normal calls, and if it encounters a throttle, it will return nil and retry with a random number between 0 and 8000ms, etc. etc. 

The advantage of using a random interval with 0 minimum is that the calls from many simultaneous clients are more smoothly distributed, and exponential backoff means that we won't get the weird throttling errors in the e2e tests anymore. 

For more on the reasoning (this PR uses a "full jitter" strategy), see [this article](https://aws.amazon.com/blogs/architecture/exponential-backoff-and-jitter/) on the AWS architecture blog. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.